requirements.txt
pynput
pyobjc-framework-Quartz 